requirements.txt
autopep8==1.6.0
click==8.1.3
glfw==2.5.3
multipledispatch==0.6.0
numpy==1.23.0
Pillow==9.1.1
pycodestyle==2.8.0
PyGLM==2.5.7
PyOpenGL==3.1.6
pyrr==0.10.3
six==1.16.0
toml==0.10.2
